Crude work-around for when (savages) clean.
Currently if you run cabal clean the configure
doesn't not get triggered on subsequent commands
because the cabal file isn't detemined to be out
of date. This just checks if dist exists and
forcibly configures in that case, this is crude
and there are probably holes in it, but covers
a common case.