Added support for .mafiaignore files to control hashing
Files in the `.mafiaignore` will not be included in a source
dependency's hash. This is useful for code generated by a Setup.hs that
may contain timestamps. Such code must be in the sdist in order for
builds to work correctly, but should be ignored for the purposes of
detecting changes to a package.

A 'mafia hash' command is added so that users can easily check whether
their package is being hashed correctly. We could potentially make it a
build failure if calling this twice gives different output.