Update mafia script to remove version lock.
This matches a change in usage that is more
common for mafia now. The original reasoning
for the lock, was so we could control version
breakage of mafia itself if required. At the
time we weren't sure how stable it could /
would be. In retrospect, the mafia API has
been able to remain very stable, and it has
actually (probably unsurprisingly) been the
unrelenting churn on hackage that has caused
the most issues. This led to a prefer the
latest mafia, and it promises to be backwards
compatiable approach. This means that later
versions of mafia can handle hackage breakage
and projects will keep building from checkout.

The script becomes a small version manager for
mafia. From checkout, everything just works
without a previous install of mafia:

  './mafia build'

This will install the latest version of mafia
and build your project. If you already have a
local build of mafia, it will continue to use
this, if you want to use the latest run:

  './mafia set-mafia-version'

This will build and link the latest mafia version.
If you want to run a specific git commit of mafia
you can:

  './mafia set-mafia-version f52d11e0beeedbd8ba31ad8ec3a760a8b670c138'

As it stands it removes the interactive upgrade
command on the local mafia script.