Continue building after failing to install build-tool
Some packages refer to build-tools that are not cabal packages.
For example, "zip-archive" depends on "unzip", but this is talking about a binary, not a cabal package.
This is unfortunate, and perhaps the package shouldn't include this as it isn't a build tool, but we should probably provide some way to continue building regardless.